refactor: centralize tool types in proto pkg and unmarshal permission params
